VS-2575 (Error Pages) Fix mount always being set to English when page resolves to hst:pagenotfound
This was due to an oversight in the site map handler for the translation fallback. The mount was changed on the request to English to lookup the document on the english mount, and had to be set back to the original locale after this check had been completed. However I was only changing the mount back to the original locale when an english fallback document was found