VS-3052 (Featured widgets) Move featured widgets to sandbox & create folder for production widgets in general/navigation/widgets
I've moved the widgets folder from content/widgets to content/general/navigation/widgets (I put them in a widgets folder as I expect the banner would be in content/general/navigation/banner). I have updated the sitemap so that the url widgets/<document> maps to /content/general/navigation/widgets/<document>

For the sandbox environment, I've moved the existing widgets into /content/sandbox/widgets

I have noticed an issue with 404 pages if the user happens to visit a url associated with a featured widget. If this is an easy fix I'll address it in a follow up commit